build-release.sh: no longer needs each dockerfile version specified explicitly

+++ build-release.sh
@@ -6,14 +6,14 @@ case $version in
   centos5|centos)
     version=centos5
     ;;
-  alpine)
-    ;;
-  *)
-    echo "version unsupported, try alpine or centos5"
-    exit 1
-    ;;
 esac
 
-docker build -f Dockerfile.$version -t $version-smell-baron . || exit 1
+dockerfile=Dockerfile.$version
+if [ ! -f $dockerfile ] ; then
+  echo "version unsupported, try alpine or centos5"
+  exit 1
+fi
+
+docker build -f $dockerfile -t $version-smell-baron . || exit 1
 docker run $version-smell-baron true || exit 1
 docker cp $(docker ps -aq | head -n1):/smell-baron/smell-baron .
